pyproject.toml
[build-system]
requires = ["maturin>=1.2,<2.0"]
build-backend = "maturin"

[project]
name = "convex"
version = "0.6.0a2"  # Also update in __init__.py
description = "Python client for the reactive backend-as-a-service Convex."
authors = [
  { name = "Convex, Inc", email = "support@convex.dev" },
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: JavaScript",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Rust",
    "Topic :: Database :: Front-Ends",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
license = "Apache-2.0"

[project.urls]
homepage = "https://convex.dev"
repository = "https://github.com/get-convex/convex-py"
documentation = "https://docs.convex.dev"

[tool.maturin]
python-source = "python"
python-packages = ["convex", "_convex"]
bindings = "pyo3"
features = ["pyo3/extension-module"]

[tool.poetry]
# We publish with maturin which uses the [project] metadata above
# but these three fields are required just to be able to use poetry.
name = "convex"
version = "0.6.0a2"  # Also update in __init__.py
description = "Python client for the reactive backend-as-a-service Convex."
authors = ["Convex, Inc. <support@convex.dev>"]

[tool.poetry.dependencies]
python = ">=3.9"
python-dotenv = "^1.0.0"
requests = "^2.28.1"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.0"
black = "^23.0.0"
flake8 = { version = "^6.0.0", python = "^3.9" }
flake8-bugbear = { version = "^23.0.0", python = "^3.9" }
flake8-docstrings = { version = "^1.7.0", python = "^3.9" }
flake8-noqa = { version = "^1.3.0", python = "^3.9" }
isort = "^5.10.1"
maturin = ">=1.2,<2.0"
mypy = "^1.0"
pytest = "^7.1.2"
pytest-profiling = "^1.7.0"
twine = "^4.0.2"
types-requests = "^2.27.30"
typing-extensions = "^4.4.0"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests",
    "integration",
    "convex"
]

[tool.mypy]
strict = true

[tool.isort]
skip = [".venv"]
